Work around C4668 bug in Windows SDK headers
There's a bug in recent versions (as of 2017-06-05) of the Windows SDK
that causes C4668 to be emitted from corecrt_math.h. Until a fix for
this is released, we disable this warning: '_M_HYBRID_X86_ARM64' is not
defined as a preprocessor macro, replacing with '0' for '#if/#elif'

Note: upstream boringssl suppresses this warning itself already, but we
need to patch the version that gRPC v1.3.4 is using.